Refatoração: Ajustes na extensibilidade e modularidade do projeto no arquivo irrf.py, criando arquivos para deduction, incomes e const
